homework_07/dev_requirements.txt
django-debug-toolbar==4.4.2
pylint
pylint_django=2.5.5
black=24.4.2